Added new vignette, colors (#51)
* Added new vignette, colors

Added new vignette to the vignettes, and the file colors.Rda in the
data folder, which is used for the color palettes

* Many changes to new Vignette

All requested edits done.
Number of external packages used  sharply reduced.
Several new examples included.
Example netcdf file include in /inst/extdata, and successfully used in
vignette.
Info file for colors.rda included (colors.R)
A lot of minor edits.

* Labelling Fixes, Improved Data Credit

Bugs in SODA and IMI titles fixed.  References to the original sources
of the data for many of the datasets improved.